Adding Docker-only installation in addition to nodejs/npm license setup
This commit add instructions and the Dockerfile.configure to setup
the license using Docker itself. This is intended to be run manually,
but at some point can be automated.

This commit adds a "configure" section that helps performing the setup
using Docker itself. Since the installer requires Node.js, this imposes
the host to have it installed. This options exercises some potential
hacks to have a docker-only setup.

*	new file:   Dockerfile.configure
- The dockerfile to run the `npm run configure`

*	modified:   README.md
- Updating the read me with instructions to copy the license file
  from the container itself to the host manually.
- TODO: Figure out a way to automate the process. This cannot be
  called from a shell scripts.